feat: Jpa 레포지토리 구현 및 다른 사람 대비 사용량 조회 메서드 구현 #23
