+++ src/main/java/com/project/expensemanage/domain/analysis/repository/dto/ExpenditureTotalUser.java
@@ -1,0 +1,13 @@
+package com.project.expensemanage.domain.analysis.repository.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+
+@Builder
+@Getter
+@AllArgsConstructor
+public class ExpenditureTotalUser {
+  private Long totalExpenditure;
+  private Long userExpenditure;
+}
